refactor: extraer custom hooks en Solicitud.jsx
