+++ frontend_uacm/src/utils/getCookie.js
@@ -1,0 +1,9 @@
+export function getCookie(name) {
+  let value = null
+  if (document.cookie)
+    document.cookie.split(';').forEach(c => {
+      const t = c.trim()
+      if (t.startsWith(name + '=')) value = decodeURIComponent(t.substring(name.length + 1))
+    })
+  return value
+}
